Don't split core by state
